Make all remote calls accept CancellationTokens
Fixes #56 by adding an optional CancellationToken parameter to every
method that ends up doing an HTTP request. Some of these seem ill-advised
(e.g. allowing the release of a semaphore to be canceled) but in many
cases they should only be used if the call can possibly fail and a
secondary timeout is needed.